Add files via upload
Index.js is the main file. Here is the ejs file (index.ejs) rendered. In the ejs file the styling is added which is the CSS file(main.css) and the logic of the todo list is in the script.js file which is also connected with the ejs file. 
The dependencies installed = body-parser, ejs, express.
ejs is an embedded javascript file where html document is written and js code also can be embedded.